Change constant expression to fit in 16-bits
Previously Kermit was manipulating some constant expressions using
MASM and the results exceeded what was allowed in 16-bits. Instead of
throwing away the extra bits, JWASM is promoting the constant to
32-bits. Of course, then the assembly instructions using that constant
break because they presume the result is a 16-bit number.

It is not entirely clear what the correct solution should be, but
doing a boolean AND with 0xFFFF does truncate the value to 16-bits, so
that's what I'm trying now.