Parenthesize expression evaluation for MASM 6.0
rdbuf+2[bx] is parsed in MASM 6 (which JWASM is compatible with) as
rdbuf+(2[bx]) which is not what anybody wants.

Kermit gets around it by using the /Zm (MASM 5.1 compat) mode, but
JWASM doesn't suppport that. Instead, I have to parenthesize it as
(rdbuf+2)[bx]

I searched for expressions that might be affected using:

	egrep -- '^[^;]*\+[^,;]+\[' *

I hope that caught them all.