Replace 'ifdef x + y + z' with series of elseifdefs
Mskermit originally used:

   IFDEF no_graphics + no_tcp + no_network

But that doesn't work with jwasm.

I've unrolled it using ELSEIFDEF, presuming the plus sign means
"boolean OR", but I'm having trouble finding the documentation for how
it was supposed to work. Could + have meant AND since that's how it's
sometimes pronounced in English?

It looks like this is meant to put a carriage return if the text " No
graphics", " No network", or " No tcp/ip" had been printed. It makes
sense then that this should be an OR.