Remove -ms (small memory model) from jwasm
Only the C code should have been compiled for the small memory model.

Using jwasm -ms was causing a link error:

    Error! E2020: size of group DGROUP exceeds 64k by 203590 bytes

I had thought the C code and assembler had to both be set to the same
memory model, but I guess not as the program now links and runs.

Question: What memory model is the default in jwasm?