Add target upx for compression; add LITESUBSYSTEMS
UPX will compress the kermit.exe from 300 KB to 152 KB. Not too
shabby! It increase load time on an IBM 5155 by four seconds, but
that's not much compared to the 22 seconds it normally takes and
completely negligible compared to 80 seconds it takes including
parsing the MSKERMIT.INI and MSCUSTOM.INI files.

LITESUBSYSTEMS is not working yet, but this is a start at using #ifdef
to remove various features. (I do not know why it doesn't just work
since I'm not adding or removing #ifdefs, just using the existing ones.)